+++ src/test/java/es/udc/fi/dc/irlab/ppc/hcomputation/TestH2Mapper.java
@@ -18,7 +18,7 @@ package es.udc.fi.dc.irlab.ppc.hcomputation;
 
 import java.io.IOException;
 
-import org.apache.hadoop.io.LongWritable;
+import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.mrunit.mapreduce.MapDriver;
 import org.apache.mahout.math.DenseVector;
 import org.apache.mahout.math.Vector;
@@ -28,20 +28,20 @@ import org.junit.Test;
 
 public class TestH2Mapper {
 
-    private MapDriver<LongWritable, VectorWritable, LongWritable, VectorWritable> mapDriver;
+    private MapDriver<IntWritable, VectorWritable, IntWritable, VectorWritable> mapDriver;
 
     @Before
     public void setup() {
-	mapDriver = new MapDriver<LongWritable, VectorWritable, LongWritable, VectorWritable>();
+	mapDriver = new MapDriver<IntWritable, VectorWritable, IntWritable, VectorWritable>();
     }
 
     @Test
     public void testMap() throws IOException {
-	LongWritable inputKey = new LongWritable(1);
+	IntWritable inputKey = new IntWritable(1);
 	Vector inputVector = new DenseVector(new double[] { 1.0, 2.0, 3.0 });
 	VectorWritable inputValue = new VectorWritable(inputVector);
 
-	LongWritable outputKey = new LongWritable(1);
+	IntWritable outputKey = new IntWritable(1);
 	Vector outputVector = new DenseVector(new double[] { 1.0, 2.0, 3.0 });
 	VectorWritable outputValue = new VectorWritable(outputVector);
